refactor:separate replication
