+++ src/services/cluster/actors/actor.rs
@@ -1,9 +1,9 @@
-use super::replication::Replication;
 use super::PeerState;
 use crate::services::cluster::command::AddPeer;
 use crate::services::cluster::command::ClusterCommand;
 use crate::services::cluster::peer::identifier::PeerIdentifier;
 use crate::services::cluster::peer::peer::Peer;
+use crate::services::cluster::replication::replication::Replication;
 use crate::services::interface::TWrite;
 use crate::services::query_io::QueryIO;
 use std::collections::BTreeMap;
