+++ src/services/cluster/actors/mod.rs
@@ -1,4 +1,3 @@
 pub(super) mod actor;
-pub(crate) use replication::PeerState;
+pub(crate) use crate::services::cluster::replication::replication::PeerState;
 pub(crate) mod listening_actor;
-pub(crate) mod replication;
