+++ src/services/cluster/command.rs
@@ -1,9 +1,7 @@
-use crate::services::cluster::actors::{
-    replication::Replication,
-    PeerState,
-};
+use crate::services::cluster::actors::PeerState;
 use crate::services::cluster::peer::identifier::{PeerAddrs, PeerIdentifier};
 use crate::services::cluster::peer::kind::PeerKind;
+use crate::services::cluster::replication::replication::Replication;
 use crate::services::query_io::QueryIO;
 use tokio::net::TcpStream;
