+++ src/services/cluster/mod.rs
@@ -4,3 +4,4 @@ pub mod manager;
 pub mod outbound;
 pub mod command;
 pub(crate) mod peer;
+pub(crate) mod replication;
